+++ app/urls.py
@@ -4,5 +4,6 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/auth/', include('users.urls')),
-    path('api/calories/', include('calories.urls'))
+    path('api/calories/', include('calories.urls')),
+    path('api/calories/limit/', include('calory_limit.urls'))
 ]
