+++ calories/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
-from .models import Calories, CaloryLimit
+from .models import Calories
 
 admin.site.register(Calories)
-admin.site.register(CaloryLimit)
