+++ calory_limit/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import CaloryLimit
+
+admin.site.register(CaloryLimit)
