+++ calory_limit/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CaloryLimitConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'calory_limit'
