+++ calory_limit/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from .models import CaloryLimit
+
+class CaloryLimitSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = CaloryLimit
+        fields = '__all__'
